+++ src/Microsoft.AspNetCore.SignalR/HubEndPoint.cs
@@ -2,18 +2,15 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Buffers;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Threading.Tasks.Channels;
 using Microsoft.AspNetCore.SignalR.Internal;
 using Microsoft.AspNetCore.SignalR.Internal.Protocol;
 using Microsoft.AspNetCore.Sockets;
-using Microsoft.AspNetCore.Sockets.Internal.Formatters;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Internal;
 using Microsoft.Extensions.Logging;
@@ -34,8 +31,6 @@ namespace Microsoft.AspNetCore.SignalR
 
     public class HubEndPoint<THub, TClient> : IInvocationBinder where THub : Hub<TClient>
     {
-        private static readonly Type[] _excludeInterfaces = new[] { typeof(Hub<>), typeof(IDisposable) };
-
         private readonly Dictionary<string, HubMethodDescriptor> _methods = new Dictionary<string, HubMethodDescriptor>(StringComparer.OrdinalIgnoreCase);
 
         private readonly HubLifetimeManager<THub> _lifetimeManager;
@@ -388,10 +383,9 @@ namespace Microsoft.AspNetCore.SignalR
         private void DiscoverHubMethods()
         {
             var hubType = typeof(THub);
+            var hubTypeInfo = hubType.GetTypeInfo();
 
-            var allInterfaceMethods = _excludeInterfaces.SelectMany(i => GetInterfaceMethods(hubType, i));
-            foreach (var methodInfo in hubType.GetMethods(BindingFlags.Public | BindingFlags.Instance)
-                .Except(allInterfaceMethods).Where(m => IsHubMethod(m)))
+            foreach (var methodInfo in HubReflectionHelper.GetHubMethods(hubType))
             {
                 var methodName = methodInfo.Name;
 
@@ -400,7 +394,7 @@ namespace Microsoft.AspNetCore.SignalR
                     throw new NotSupportedException($"Duplicate definitions of '{methodName}'. Overloading is not supported.");
                 }
 
-                var executor = ObjectMethodExecutor.Create(methodInfo, hubType.GetTypeInfo());
+                var executor = ObjectMethodExecutor.Create(methodInfo, hubTypeInfo);
                 _methods[methodName] = new HubMethodDescriptor(executor);
 
                 if (_logger.IsEnabled(LogLevel.Debug))
@@ -408,34 +402,6 @@ namespace Microsoft.AspNetCore.SignalR
                     _logger.LogDebug("Hub method '{methodName}' is bound", methodName);
                 }
             }
-        }
-
-        private static bool IsHubMethod(MethodInfo methodInfo)
-        {
-            var baseDefinition = methodInfo.GetBaseDefinition().DeclaringType;
-            if (typeof(object) == baseDefinition || methodInfo.IsSpecialName)
-            {
-                return false;
-            }
-
-            // removes methods such as Hub<TClient>.OnConnectedAsync
-            var baseType = baseDefinition.GetTypeInfo().IsGenericType ? baseDefinition.GetGenericTypeDefinition() : baseDefinition;
-            if (typeof(Hub<>) == baseType)
-            {
-                return false;
-            }
-
-            return true;
-        }
-
-        private static IEnumerable<MethodInfo> GetInterfaceMethods(Type type, Type iface)
-        {
-            if (!iface.IsAssignableFrom(type))
-            {
-                return Enumerable.Empty<MethodInfo>();
-            }
-
-            return type.GetInterfaceMap(iface).TargetMethods;
         }
 
         Type IInvocationBinder.GetReturnType(string invocationId)
